+++ tests/testthat/test-docinfo.R
@@ -144,28 +144,3 @@ test_that("conversion to/from docinfo()", {
     d3 <- as_docinfo(as_xmp(x2))
     expect_equal(d3$subject, "Generic Subject")
 })
-
-test_that("from_date_pdfmark()", {
-    expect_null(to_date_pdfmark(NULL))
-    expect_error(from_date_pdfmark("D:20081"), "Couldn't parse pdfmark date")
-    expect_equal(format(from_date_pdfmark("D:20081206"), format = "%Y-%m-%d"),
-                 "2008-12-06")
-    expect_equal(format(from_date_pdfmark("20081206"), format = "%Y-%m-%d"),
-                 "2008-12-06")
-    expect_equal(format(from_date_pdfmark("D:2008"), format = "%Y-%m-%d"),
-                 "2008-01-01")
-    expect_equal(format(from_date_pdfmark("D:200802"), format = "%Y-%m-%d"),
-                 "2008-02-01")
-    expect_equal(format(from_date_pdfmark("D:20080210"), format = "%Y-%m-%d"),
-                 "2008-02-10")
-    expect_equal(format(from_date_pdfmark("D:2008021020"), format = "%Y-%m-%d %H:%M:%S"),
-                 "2008-02-10 20:00:00")
-    expect_equal(format(from_date_pdfmark("D:200802102010"), format = "%Y-%m-%d %H:%M:%S"),
-                 "2008-02-10 20:10:00")
-    expect_equal(format(from_date_pdfmark("D:20080210201005"), format = "%Y-%m-%d %H:%M:%S"),
-                 "2008-02-10 20:10:05")
-    expect_equal(format(from_date_pdfmark("D:20080210201005+05"), format = "%Y-%m-%d %H:%M:%S%z"),
-                 "2008-02-10 15:10:05+0000")
-    expect_equal(format(from_date_pdfmark("D:20080210201005+0500"), format = "%Y-%m-%d %H:%M:%S%z"),
-                 "2008-02-10 15:10:05+0000")
-})